Cleanups of routes.js
